+++ lib/interface-types/src/interpreter/instructions/memory_to_string.rs
@@ -1,53 +1,63 @@
-use crate::interpreter::wasm::values::InterfaceValue;
-use std::{cell::Cell, convert::TryFrom};
+use super::to_native;
+use crate::{
+    errors::{InstructionError, InstructionErrorKind},
+    interpreter::{wasm::values::InterfaceValue, Instruction},
+};
+use std::cell::Cell;
 
 executable_instruction!(
-    memory_to_string(instruction_name: String) -> _ {
+    memory_to_string(instruction: Instruction) -> _ {
         move |runtime| -> _ {
-            match runtime.stack.pop(2) {
-                Some(inputs) => match runtime.wasm_instance.memory(0) {
-                    Some(memory) => {
-                        let length = i32::try_from(&inputs[0])? as usize;
-                        let pointer = i32::try_from(&inputs[1])? as usize;
-                        let memory_view = memory.view();
+            let inputs = runtime.stack.pop(2).ok_or_else(|| {
+                InstructionError::new(
+                    instruction,
+                    InstructionErrorKind::StackIsTooSmall { needed: 2 },
+                )
+            })?;
 
-                        if memory_view.len() < pointer + length {
-                            return Err(format!(
-                                "`{}` failed because it has to read out of the memory bounds (index {} > memory length {}).",
-                                instruction_name,
-                                pointer + length,
-                                memory_view.len()
-                            ));
-                        }
+            let memory_index: u32 = 0;
 
-                        let data: Vec<u8> = (&memory_view[pointer..pointer + length])
-                            .iter()
-                            .map(Cell::get)
-                            .collect();
+            let memory = runtime
+                .wasm_instance
+                .memory(memory_index as usize)
+                .ok_or_else(|| {
+                    InstructionError::new(
+                        instruction,
+                        InstructionErrorKind::MemoryIsMissing { memory_index },
+                    )
+                })?;
 
-                        match String::from_utf8(data) {
-                            Ok(string) => {
-                                runtime.stack.push(InterfaceValue::String(string));
+            let length = to_native::<i32>(&inputs[0], instruction)? as usize;
+            let pointer = to_native::<i32>(&inputs[1], instruction)? as usize;
+            let memory_view = memory.view();
 
-                                Ok(())
-                            }
-                            Err(utf8_error) => Err(format!(
-                                "`{}` failed because the read string isn't UTF-8 valid ({}).",
-                                instruction_name,
-                                utf8_error,
-                            ))
-                        }
-                    }
-                    None => Err(format!(
-                        "`{}` failed because there is no memory to read.",
-                        instruction_name
-                    ))
-                }
-                None => Err(format!(
-                    "`{}` failed because there is not enough data on the stack (needs 2).",
-                    instruction_name,
-                ))
+            if length == 0 {
+                runtime.stack.push(InterfaceValue::String("".into()));
+
+                return Ok(())
             }
+
+            if memory_view.len() <= pointer + length - 1 {
+                return Err(InstructionError::new(
+                    instruction,
+                    InstructionErrorKind::MemoryOutOfBoundsAccess {
+                        index: pointer + length,
+                        length: memory_view.len(),
+                    },
+                ));
+            }
+
+            let data: Vec<u8> = (&memory_view[pointer..=pointer + length - 1])
+                .iter()
+                .map(Cell::get)
+                .collect();
+
+            let string = String::from_utf8(data)
+                .map_err(|error| InstructionError::new(instruction, InstructionErrorKind::String(error)))?;
+
+            runtime.stack.push(InterfaceValue::String(string));
+
+            Ok(())
         }
     }
 );
@@ -75,6 +85,24 @@ mod tests {
     );
 
     test_executable_instruction!(
+        test_memory_to_string__empty_string =
+            instructions: [
+                Instruction::ArgumentGet { index: 1 },
+                Instruction::ArgumentGet { index: 0 },
+                Instruction::MemoryToString,
+            ],
+            invocation_inputs: [
+                InterfaceValue::I32(0),
+                InterfaceValue::I32(0),
+            ],
+            instance: Instance {
+                memory: Memory::new(vec![]),
+                ..Default::default()
+            },
+            stack: [InterfaceValue::String("".into())],
+    );
+
+    test_executable_instruction!(
         test_memory_to_string__read_out_of_memory =
             instructions: [
                 Instruction::ArgumentGet { index: 1 },
@@ -91,7 +119,7 @@ mod tests {
                 memory: Memory::new("Hello!".as_bytes().iter().map(|u| Cell::new(*u)).collect()),
                 ..Default::default()
             },
-            error: r#"`memory-to-string` failed because it has to read out of the memory bounds (index 13 > memory length 6)."#,
+            error: r#"`memory-to-string` read out of the memory bounds (index 13 > memory length 6)"#,
     );
 
     test_executable_instruction!(
@@ -111,7 +139,7 @@ mod tests {
                 memory: Memory::new(vec![0, 159, 146, 150].iter().map(|b| Cell::new(*b)).collect::<Vec<Cell<u8>>>()),
                 ..Default::default()
             },
-            error: r#"`memory-to-string` failed because the read string isn't UTF-8 valid (invalid utf-8 sequence of 1 bytes from index 1)."#,
+            error: r#"`memory-to-string` invalid utf-8 sequence of 1 bytes from index 1"#,
     );
 
     test_executable_instruction!(
@@ -126,6 +154,6 @@ mod tests {
                 InterfaceValue::I32(0),
             ],
             instance: Instance::new(),
-            error: r#"`memory-to-string` failed because there is not enough data on the stack (needs 2)."#,
+            error: r#"`memory-to-string` needed to read `2` value(s) from the stack, but it doesn't contain enough data"#,
     );
 }